bot/requirements.txt
absl-py==0.13.0
schedule==1.1.0
selenium==3.141.0
six==1.16.0
urllib3==1.26.6
